# v1.2.0 Enhancement: Introduce optional LRU cache for massive performance improvement (10x)