--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+*.gem
+Gemfile.lock