--- a/README.md
+++ b/README.md
@@ -1,3 +1,7 @@
+# ARCHIVED
+
+I am not in a position to maintain this as frequently as I would have liked, and my logstash work required a lot of updating for Logstash 8, so I've made the decision to archive this project. It was last known to work with Logstash 7. Please look to Logstash's own version of this plugin (which this was forked from). 
+
 # Logstash Plugin
 
 [![Travis Build Status](https://travis-ci.com/logstash-plugins/logstash-filter-memcached.svg)](https://travis-ci.com/logstash-plugins/logstash-filter-memcached)