--- a/logstash-filter-memcached.gemspec
+++ b/logstash-filter-memcached.gemspec
@@ -1,6 +1,6 @@
 Gem::Specification.new do |s|
   s.name          = 'logstash-filter-memcached'
-  s.version       = '1.1.0'
+  s.version       = '1.2.0'
   s.licenses      = ['Apache-2.0']
   s.summary       = 'A Logstash filter plugin for interacting with memcached'
   s.authors       = ["Elastic"]