--- a/qa/logstash.conf
+++ b/qa/logstash.conf
@@ -41,8 +41,8 @@ filter {
     get => {
       "%{message}" => "ipinfo"
     }
-    # lru_cache_ttl => 60
-    # lru_cache_max_size => 1000
+    lru_cache_ttl => 60
+    lru_cache_max_size => 1000
   }
 
   metrics {