--- a/qa/logstash.conf
+++ b/qa/logstash.conf
@@ -42,7 +42,7 @@ filter {
       "%{message}" => "ipinfo"
     }
     lru_cache_ttl => 60
-    lru_cache_max_size => 1000
+    lru_cache_max_entries => 1000
   }
 
   metrics {