--- a/qa/process-portion.conf
+++ b/qa/process-portion.conf
@@ -1,0 +1,59 @@
+# To setup:
+# 
+# For TCP memcached on 127.0.0.1:11211
+#
+# memcached -d -l 127.0.0.1
+# for i in `seq 1 10`
+# do
+#   echo -en 'set perf:10.0.0.'$i' 0 10000 100\r\n1234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890\r\n' | nc 127.0.0.1 11211
+# done
+
+input {
+  generator {
+    lines => [
+      "10.0.0.1",
+      "10.0.0.2",
+      "10.0.0.3",
+      "10.0.0.4",
+      "10.0.0.5",
+      "10.0.0.6",
+      "10.0.0.7",
+      "10.0.0.8",
+      "10.0.0.9",
+      "10.0.0.10"
+    ]
+  }
+}
+
+filter {
+
+  throttle {
+    before_count => 1000
+    period => 10
+    max_age => 20
+    key => "any"
+    add_tag => "pass_to_memcache"
+  }
+
+  if "pass_to_memcache" in [tags] {
+    memcached {
+      hosts => ["127.0.0.1:11211"]
+      namespace => "perf"
+      id => "memcached-get"
+      get => {
+        "%{message}" => "ipinfo"
+      }
+    }
+  }
+
+  # drop {}
+}
+
+output {
+  stdout { codec => "json_lines" }
+}
+
+# In another window...
+#
+# curl -s 127.0.0.1:9602/_node/stats/pipelines/main | jq '.pipelines.main.plugins.filters[] | select(.id == "memcached-get") | (.events.duration_in_millis / .events.out)'
+#