@cameronkerrnz 2021-06-09
I've had this in production for much of a day so far, including at least half of day during the peak usage period. Here's what it looks like from a metrics monitoring point-of-view (this data sourced using logstash_exporter)

![image](https://user-images.githubusercontent.com/938702/121277555-f6b51400-c924-11eb-9c64-65edffdf9e51.png)

The blue line shows when I activated the change to use the LRU cache. On the top-graph, the green line shows the sum total of time spent per second executing this instance of the memcached filter. (so it could in theory reach a max of num_pipeline_workers). The yellow line (ignore the red dots; those relate to non-matches in other modules) shows the number of events flowing through the pipeline.

In terms of how this relates to the performance of other things going through the pipeline, it has dropped off the top-5 slowest plugins [text target is improving the performance of a particular grok filter].